Refactor ServiceImportAggregator methods to ServiceImportController
...since the ServiceImportAggregator is no longer reused. Also refactor
aggregate SI creation code from ServiceImportController Distribute method
to createOrUpdateAggregate.
